backend_1/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib[bcrypt]
youtube-transcript-api
openai
fastapi
python-multipart
pydantic
